Feat : 설계 수정
